Backend/feat/praman-testing

+++ Backend/app/rider-platform/rider-app/Main/src-read-only/Storage/Beam/FRFSQuote.hs
@@ -15,7 +15,7 @@ import Tools.Beam.UtilsTH
 
 data FRFSQuoteT f = FRFSQuoteT
   { _type :: B.C f Domain.Types.FRFSQuote.FRFSQuoteType,
-    bppDelayedInterest :: B.C f (Kernel.Prelude.Maybe Kernel.Prelude.Int),
+    bppDelayedInterest :: B.C f (Kernel.Prelude.Maybe Kernel.Prelude.Double),
     bppItemId :: B.C f Kernel.Prelude.Text,
     bppSubscriberId :: B.C f Kernel.Prelude.Text,
     bppSubscriberUrl :: B.C f Kernel.Prelude.Text,
